Moving PointOnCurve and Scalar32Bytes to separate files.
